support rust v1.0.0-alpha

+++ src/lib.rs
@@ -2,15 +2,16 @@ extern crate time;
 use std::rand;
 
 static EPOCH:f64 = 946702800.0;
-static TIMESTAMP_SHIFT:uint = 23;
+static TIMESTAMP_SHIFT:usize = 23;
 static RANDOM_MAX_VALUE:u32 = 4194303;
 
-#[deriving(Copy)]
+#[derive(Copy)]
 pub struct Id {
     pub timestamp: f64,
     pub random_bits: u32,
 }
 
+#[allow(unstable)]
 pub fn new() -> u64 {
     let now = time_as_float(time::get_time());
     let timestamp = ((now - EPOCH) * 1000.0) as u64;
